Add functions to asynchronously load soft pointers
